+++ packages/google-cloud-networkconnectivity/src/v1beta/transport_manager_client.ts
@@ -466,6 +466,7 @@ export class TransportManagerClient {
     const transportManagerStubMethods = [
       'listRemoteTransportProfiles',
       'getRemoteTransportProfile',
+      'parseFromActivationKey',
       'listTransports',
       'getTransport',
       'getStatus',
@@ -733,6 +734,151 @@ export class TransportManagerClient {
       });
   }
   /**
+   * Gets details of a single RemoteTransportProfile given an activation key.
+   *
+   * @param {Object} request
+   *   The request object that will be sent.
+   * @param {string} request.parent
+   *   Required. Parent value for ParseFromActivationKeyRequest.
+   * @param {string} request.activationKey
+   *   Required. The activation key to get the RemoteTransportProfile for.
+   * @param {object} [options]
+   *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
+   * @returns {Promise} - The promise which resolves to an array.
+   *   The first element of the array is an object representing {@link protos.google.cloud.networkconnectivity.v1beta.ParseFromActivationKeyResponse|ParseFromActivationKeyResponse}.
+   *   Please see the {@link https://github.com/googleapis/gax-nodejs/blob/master/client-libraries.md#regular-methods | documentation }
+   *   for more details and examples.
+   * @example <caption>include:samples/generated/v1beta/transport_manager.parse_from_activation_key.js</caption>
+   * region_tag:networkconnectivity_v1beta_generated_TransportManager_ParseFromActivationKey_async
+   */
+  parseFromActivationKey(
+    request?: protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest,
+    options?: CallOptions,
+  ): Promise<
+    [
+      protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyResponse,
+      (
+        | protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest
+        | undefined
+      ),
+      {} | undefined,
+    ]
+  >;
+  parseFromActivationKey(
+    request: protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest,
+    options: CallOptions,
+    callback: Callback<
+      protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyResponse,
+      | protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest
+      | null
+      | undefined,
+      {} | null | undefined
+    >,
+  ): void;
+  parseFromActivationKey(
+    request: protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest,
+    callback: Callback<
+      protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyResponse,
+      | protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest
+      | null
+      | undefined,
+      {} | null | undefined
+    >,
+  ): void;
+  parseFromActivationKey(
+    request?: protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest,
+    optionsOrCallback?:
+      | CallOptions
+      | Callback<
+          protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyResponse,
+          | protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest
+          | null
+          | undefined,
+          {} | null | undefined
+        >,
+    callback?: Callback<
+      protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyResponse,
+      | protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest
+      | null
+      | undefined,
+      {} | null | undefined
+    >,
+  ): Promise<
+    [
+      protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyResponse,
+      (
+        | protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest
+        | undefined
+      ),
+      {} | undefined,
+    ]
+  > | void {
+    request = request || {};
+    let options: CallOptions;
+    if (typeof optionsOrCallback === 'function' && callback === undefined) {
+      callback = optionsOrCallback;
+      options = {};
+    } else {
+      options = optionsOrCallback as CallOptions;
+    }
+    options = options || {};
+    options.otherArgs = options.otherArgs || {};
+    options.otherArgs.headers = options.otherArgs.headers || {};
+    options.otherArgs.headers['x-goog-request-params'] =
+      this._gaxModule.routingHeader.fromParams({
+        parent: request.parent ?? '',
+      });
+    this.initialize().catch((err) => {
+      throw err;
+    });
+    this._log.info('parseFromActivationKey request %j', request);
+    const wrappedCallback:
+      | Callback<
+          protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyResponse,
+          | protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest
+          | null
+          | undefined,
+          {} | null | undefined
+        >
+      | undefined = callback
+      ? (error, response, options, rawResponse) => {
+          this._log.info('parseFromActivationKey response %j', response);
+          callback!(error, response, options, rawResponse); // We verified callback above.
+        }
+      : undefined;
+    return this.innerApiCalls
+      .parseFromActivationKey(request, options, wrappedCallback)
+      ?.then(
+        ([response, options, rawResponse]: [
+          protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyResponse,
+          (
+            | protos.google.cloud.networkconnectivity.v1beta.IParseFromActivationKeyRequest
+            | undefined
+          ),
+          {} | undefined,
+        ]) => {
+          this._log.info('parseFromActivationKey response %j', response);
+          return [response, options, rawResponse];
+        },
+      )
+      .catch((error: any) => {
+        if (
+          error &&
+          'statusDetails' in error &&
+          error.statusDetails instanceof Array
+        ) {
+          const protos = this._gaxModule.protobuf.Root.fromJSON(
+            jsonProtos,
+          ) as unknown as gax.protobuf.Type;
+          error.statusDetails = decodeAnyProtosInArray(
+            error.statusDetails,
+            protos,
+          );
+        }
+        throw error;
+      });
+  }
+  /**
    * Gets details of a single Transport.
    *
    * @param {Object} request
@@ -882,6 +1028,9 @@ export class TransportManagerClient {
    *   The request object that will be sent.
    * @param {string} request.name
    *   Required. Name of the resource.
+   * @param {boolean} [request.skipCache]
+   *   Optional. If set to true, the response will bypass any caches and return
+   *   the freshest possible data.
    * @param {object} [options]
    *   Call options. See {@link https://googleapis.dev/nodejs/google-gax/latest/interfaces/CallOptions.html|CallOptions} for more details.
    * @returns {Promise} - The promise which resolves to an array.