+++ packages/google-cloud-spanner-api/webpack.config.js
@@ -17,8 +17,8 @@ const path = require('path');
 module.exports = {
   entry: './src/index.ts',
   output: {
-    library: 'SpannerApi',
-    filename: './spanner-api.js',
+    library: 'Spanner',
+    filename: './spanner.js',
   },
   node: {
     child_process: 'empty',