+++ core/packages/gax/CHANGELOG.md
@@ -6,6 +6,24 @@
 [1]: https://www.npmjs.com/package/gax-nodejs?activeTab=versions
 
 
+## [6.0.1](https://github.com/googleapis/google-cloud-node/compare/google-gax-v6.0.0...google-gax-v6.0.1) (2026-08-10)
+
+
+### Bug Fixes
+
+* Upgrade all node core library dependencies to their latest node 22 version ([#9084](https://github.com/googleapis/google-cloud-node/issues/9084)) ([af8074c](https://github.com/googleapis/google-cloud-node/commit/af8074c0281f4f8dc40570f580f4cdebf1d72cde))
+
+## [6.0.0](https://github.com/googleapis/google-cloud-node/compare/google-gax-v5.0.8...google-gax-v6.0.0) (2026-07-29)
+
+
+### ⚠ BREAKING CHANGES
+
+* Update gax to minimum Node version of 22. ([#8989](https://github.com/googleapis/google-cloud-node/issues/8989))
+
+### Features
+
+* Update gax to minimum Node version of 22. ([#8989](https://github.com/googleapis/google-cloud-node/issues/8989)) ([792ffaf](https://github.com/googleapis/google-cloud-node/commit/792ffaf7389e601b906ad9870707f89326af4257)), refs [#8985](https://github.com/googleapis/google-cloud-node/issues/8985)
+
 ## [5.0.8](https://github.com/googleapis/google-cloud-node/compare/google-gax-v5.0.7...google-gax-v5.0.8) (2026-07-23)