+++ packages/grafeas/CHANGELOG.md
@@ -4,6 +4,24 @@
 
 [1]: https://www.npmjs.com/package/@google-cloud/grafeas?activeTab=versions
 
+## [7.0.0](https://github.com/googleapis/google-cloud-node/compare/grafeas-v6.3.0...grafeas-v7.0.0) (2026-08-04)
+
+
+### ⚠ BREAKING CHANGES
+
+* upgrade minimum Node.js version to 22 ([#9026](https://github.com/googleapis/google-cloud-node/issues/9026))
+
+### Features
+
+* Upgrade minimum Node.js version to 22 ([#9026](https://github.com/googleapis/google-cloud-node/issues/9026)) ([1ed703f](https://github.com/googleapis/google-cloud-node/commit/1ed703f36743d61fb302b42a6cee53c07034b2fe))
+
+## [6.3.0](https://github.com/googleapis/google-cloud-node/compare/grafeas-v6.2.0...grafeas-v6.3.0) (2026-07-29)
+
+
+### Features
+
+* Update API sources and regenerate ([#8999](https://github.com/googleapis/google-cloud-node/issues/8999)) ([5c36365](https://github.com/googleapis/google-cloud-node/commit/5c363650ee4de08a81b1856e16693bdd8ad99ca9))
+
 ## [6.2.0](https://github.com/googleapis/google-cloud-node/compare/grafeas-v6.1.2...grafeas-v6.2.0) (2026-06-16)